Interrupt,HPET Timer and Local APIC Timer.

+++ include/lib/string.h
@@ -2,6 +2,6 @@
 
 void *memcpy(void *to,const void *from,int n);
 
-char *itoa(long val, char *buf, unsigned int radix,
+char *itoa(long long val, char *buf, unsigned int radix,
    char alignType,char alignChar,char isUnsigned);
 int strlen(const char *string);
